Move more inner classes to top-level classes
