+++ polaris-core/src/main/java/org/apache/polaris/core/auth/PolarisGrantManager.java
@@ -18,20 +18,15 @@
  */
 package org.apache.polaris.core.auth;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.annotation.Nonnull;
 import jakarta.annotation.Nullable;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import org.apache.polaris.core.PolarisCallContext;
 import org.apache.polaris.core.entity.PolarisBaseEntity;
 import org.apache.polaris.core.entity.PolarisEntityCore;
-import org.apache.polaris.core.entity.PolarisGrantRecord;
 import org.apache.polaris.core.entity.PolarisPrivilege;
-import org.apache.polaris.core.persistence.dao.entity.BaseResult;
+import org.apache.polaris.core.persistence.dao.entity.LoadGrantsResult;
+import org.apache.polaris.core.persistence.dao.entity.PrivilegeResult;
 
 /** Manage grants for Polaris entities. */
 public interface PolarisGrantManager {
@@ -170,136 +165,4 @@ public interface PolarisGrantManager {
   @Nonnull
   LoadGrantsResult loadGrantsToGrantee(
       PolarisCallContext callCtx, long granteeCatalogId, long granteeId);
-
-  /** Result of a grant/revoke privilege call */
-  class PrivilegeResult extends BaseResult {
-
-    // null if not success.
-    private final PolarisGrantRecord grantRecord;
-
-    /**
-     * Constructor for an error
-     *
-     * @param errorCode error code, cannot be SUCCESS
-     * @param extraInformation extra information
-     */
-    public PrivilegeResult(
-        @Nonnull BaseResult.ReturnStatus errorCode, @Nullable String extraInformation) {
-      super(errorCode, extraInformation);
-      this.grantRecord = null;
-    }
-
-    /**
-     * Constructor for success
-     *
-     * @param grantRecord grant record being granted or revoked
-     */
-    public PrivilegeResult(@Nonnull PolarisGrantRecord grantRecord) {
-      super(BaseResult.ReturnStatus.SUCCESS);
-      this.grantRecord = grantRecord;
-    }
-
-    @JsonCreator
-    private PrivilegeResult(
-        @JsonProperty("returnStatus") @Nonnull BaseResult.ReturnStatus returnStatus,
-        @JsonProperty("extraInformation") String extraInformation,
-        @JsonProperty("grantRecord") PolarisGrantRecord grantRecord) {
-      super(returnStatus, extraInformation);
-      this.grantRecord = grantRecord;
-    }
-
-    public PolarisGrantRecord getGrantRecord() {
-      return grantRecord;
-    }
-  }
-
-  /** Result of a load grants call */
-  class LoadGrantsResult extends BaseResult {
-    // true if success. If false, the caller should retry because of some concurrent change
-    private final int grantsVersion;
-
-    // null if not success. Else set of grants records on a securable or to a grantee
-    private final List<PolarisGrantRecord> grantRecords;
-
-    // null if not success. Else, for each grant record, list of securable or grantee entities
-    private final List<PolarisBaseEntity> entities;
-
-    /**
-     * Constructor for an error
-     *
-     * @param errorCode error code, cannot be SUCCESS
-     * @param extraInformation extra information
-     */
-    public LoadGrantsResult(
-        @Nonnull BaseResult.ReturnStatus errorCode, @Nullable String extraInformation) {
-      super(errorCode, extraInformation);
-      this.grantsVersion = 0;
-      this.grantRecords = null;
-      this.entities = null;
-    }
-
-    /**
-     * Constructor for success
-     *
-     * @param grantsVersion version of the grants
-     * @param grantRecords set of grant records
-     */
-    public LoadGrantsResult(
-        int grantsVersion,
-        @Nonnull List<PolarisGrantRecord> grantRecords,
-        List<PolarisBaseEntity> entities) {
-      super(BaseResult.ReturnStatus.SUCCESS);
-      this.grantsVersion = grantsVersion;
-      this.grantRecords = grantRecords;
-      this.entities = entities;
-    }
-
-    @JsonCreator
-    private LoadGrantsResult(
-        @JsonProperty("returnStatus") @Nonnull BaseResult.ReturnStatus returnStatus,
-        @JsonProperty("extraInformation") String extraInformation,
-        @JsonProperty("grantsVersion") int grantsVersion,
-        @JsonProperty("grantRecords") List<PolarisGrantRecord> grantRecords,
-        @JsonProperty("entities") List<PolarisBaseEntity> entities) {
-      super(returnStatus, extraInformation);
-      this.grantsVersion = grantsVersion;
-      this.grantRecords = grantRecords;
-      // old GS code might not serialize this argument
-      this.entities = entities;
-    }
-
-    public int getGrantsVersion() {
-      return grantsVersion;
-    }
-
-    public List<PolarisGrantRecord> getGrantRecords() {
-      return grantRecords;
-    }
-
-    public List<PolarisBaseEntity> getEntities() {
-      return entities;
-    }
-
-    @JsonIgnore
-    public Map<Long, PolarisBaseEntity> getEntitiesAsMap() {
-      return (this.getEntities() == null)
-          ? null
-          : this.getEntities().stream()
-              .collect(Collectors.toMap(PolarisBaseEntity::getId, entity -> entity));
-    }
-
-    @Override
-    public String toString() {
-      return "LoadGrantsResult{"
-          + "grantsVersion="
-          + grantsVersion
-          + ", grantRecords="
-          + grantRecords
-          + ", entities="
-          + entities
-          + ", returnStatus="
-          + getReturnStatus()
-          + '}';
-    }
-  }
 }
